Reduce memory footprint of core parser types
Memory optimizations:
- Span: use u32 for start/end (16 -> 8 bytes, max 4GB files)
- Node: box anchor/tag into Properties (96 -> 48 bytes)
- ErrorKind: use u16 for indentation context (24 -> 8 bytes)
- ParseError: remove unused expected/found fields (48 -> 24 bytes)

Documentation:
- Update ARCHITECTURE.md with memory optimization trade-offs
- Document limitations (4GB file size, 65k indentation)
- Remove ignored doctests, add test_zero_copy_parsing

All 77 tests pass.

+++ rust/yaml-parser/tests/test_suite.rs
@@ -675,7 +675,7 @@ fn run_combined_error_test(error_cases: &[(String, String, String)], order_name:
     for error in &errors {
         let span = &error.span;
         assert!(
-            span.end <= combined_input.len(),
+            (span.end as usize) <= combined_input.len(),
             "Error span {span:?} exceeds input length {}",
             combined_input.len()
         );
